refactor: add build script for html example chore: update html example readme refactor: removed duplicate examples
