+++ packages/viewport-animate-example.html/README.md
@@ -1,1 +1,3 @@
 # Viewport animate with no frameworks
+
+> Run `rush build` to get update the shared examples
