+++ packages/viewport-animate-example.html/index.html
@@ -26,7 +26,7 @@
       <link rel="stylesheet" href="no-js-styles.css" />
     </noscript>
     <script src="../viewport-animate/dist/viewport-animate.umd.js"></script>
-    <script src="../common/examples.js"></script>
+    <script src="./examples.js"></script>
     <script src="./index.js"></script>
   </head>
   <body>
