+++ packages/viewport-animate-example.html/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "viewport-animate-example.html",
+  "version": "0.1.3",
+  "description": "Plain HMTL example for viewport-animate",
+  "scripts": {
+    "test": "echo No tests found",
+    "build": "printf \"window.EXAMPLES = \" | cat - ../common/examples.json > examples.js"
+  }
+}
