+++ packages/viewport-animate-example.react/src/components/Table.tsx
@@ -1,5 +1,5 @@
-import type { Example } from "../../../common/examples.ts";
-import examplesData from "../../../common/examples.ts";
+import type { Example, ExamplesData } from "../../../common/types";
+import examplesData from "../../../common/examples.json";
 
 import Row from "./Row.tsx";
 
@@ -8,9 +8,9 @@ type Props = {
 };
 
 function Table({ example }: Props) {
-  const DEFAULTS = examplesData["defaults"];
-  const displayTexts = examplesData["display-texts"];
-  const displayTextsKeys = Object.keys(examplesData['display-texts']);
+  const DEFAULTS: ExamplesData["defaults"] = examplesData["defaults"];
+  const displayTexts: ExamplesData["display-texts"] = examplesData["display-texts"];
+  const displayTextsKeys = Object.keys(displayTexts);
   
   return (
     <div className="table-container">
