+++ packages/viewport-animate/package.json
@@ -1,6 +1,6 @@
 {
   "name": "viewport-animate",
-  "version": "0.1.3",
+  "version": "0.1.2",
   "description": "Easy and performant API for applying CSS animations when elements are scrolled into the viewport.",
   "type": "module",
   "main": "./dist/viewport-animate.umd.js",
